requirements.txt
# Core ML and NLP dependencies
torch>=2.0.0
transformers>=4.30.0
tokenizers>=0.13.0
accelerate>=0.20.0
peft>=0.4.0  # For LoRA adapters
datasets>=3.6.0,<4

# API and web framework
fastapi>=0.128.0
uvicorn[standard]>=0.40.0
uvicorn==0.40.0
pydantic>=2.12.5

# Development and evaluation
pytest>=7.0.0
pytest-asyncio>=0.21.0
hypothesis>=6.75.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.4.0
matplotlib>=3.10.8

# Utilities
pyyaml>=6.0
click>=8.0.0
tqdm>=4.65.0
numpy>=1.24.0

# Jupyter and notebook support
jupyter>=1.0.0
jupyterlab>=4.5.1
ipywidgets>=8.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
